requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
stripe==7.8.0
httpx==0.26.0
jinja2==3.1.2
email-validator==2.1.0
anthropic==0.17.0
